Checkbox Working
The Checkbuttons work because the buttons that were checking the values did something to make them correct. Instead of pulling my hair out some more I created a button to check every checkbutton value, but I did not grid them so you cannot see them.  I know this is not best practice, but I want to finish the database in time for it to be used, so I am doing this for now. Hopefully I will be able to find a real solution at some point.